fix: keep editor management in cms users module
