+++ backend/src/app.ts
@@ -20,7 +20,6 @@ import hallsRoutes from './modules/halls/halls.routes.js'
 import spacesRoutes from './modules/spaces/spaces.routes.js'
 import mediaRoutes from './modules/media/media.routes.js'
 import authRoutes from './modules/auth/auth.routes.js'
-import editorsRoutes from './modules/editors/editors.routes.js'
 import cmsUsersRoutes from './modules/cms-users/cms-users.routes.js'
 import uitdatabankRoutes from './modules/uitdatabank/uitdatabank.routes.js'
 import blogsRoutes from './modules/blogs/blogs.routes.js'
@@ -71,7 +70,6 @@ export async function buildApp(opts = {}): Promise<FastifyInstance> {
     await app.register(searchRoutes, { prefix: '/api/v1/archive/search' })
     await app.register(dashboardRoutes, { prefix: '/api/v1/dashboard' })
     await app.register(authRoutes, { prefix: '/api/v1/auth' })
-    await app.register(editorsRoutes, { prefix: '/api/v1/editors' })
     await app.register(cmsUsersRoutes, { prefix: '/api/v1/cms-users' })
 
     return app
