+++ backend/src/modules/cms-users/cms-users.schema.ts
@@ -26,16 +26,14 @@ export const cmsUserSchema = z.object({
 export const cmsUserListSchema = createPaginatedResponseSchema(cmsUserSchema)
 export const singleCmsUserSchema = createSingleResponseSchema(cmsUserSchema)
 
-export const createCmsUserSchema = z.object({
+export const createEditorSchema = z.object({
     username: z.string().min(1).transform((value) => value.trim().toLowerCase()),
     password: z.string().min(6),
-    role: z.enum(['ADMIN', 'EDITOR']).default('EDITOR'),
 })
 
-export const updateCmsUserSchema = z.object({
+export const updateEditorSchema = z.object({
     username: z.string().min(1).transform((value) => value.trim().toLowerCase()).optional(),
     password: z.string().min(6).optional(),
-    role: z.enum(['ADMIN', 'EDITOR']).optional(),
 })
 
 export const cmsUserIdSchema = z.object({
@@ -49,6 +47,6 @@ export const errorSchema = z.object({
 export type CmsUserPaginationQuery = z.infer<typeof cmsUserPaginationQuerySchema>
 export type CmsUserResponse = z.infer<typeof cmsUserSchema>
 export type CmsUserListResponse = z.infer<typeof cmsUserListSchema>
-export type CreateCmsUserInput = z.infer<typeof createCmsUserSchema>
-export type UpdateCmsUserInput = z.infer<typeof updateCmsUserSchema>
+export type CreateEditorInput = z.infer<typeof createEditorSchema>
+export type UpdateEditorInput = z.infer<typeof updateEditorSchema>
 export type CmsUserIdParams = z.infer<typeof cmsUserIdSchema>
